refactor(use_cases): REORG-1 — split received/ and query/ sub-packages
Create vultron/core/use_cases/received/ for all inbound message handler
use cases (actor, case, case_participant, embargo, note, report, status,
unknown) and vultron/core/use_cases/query/ for query use cases
(action_rules), per UC-ORG-01-001 through UC-ORG-01-003.

- git mv all 8 received-handler modules to received/
- git mv action_rules.py to query/
- Keep _helpers.py at root (shared by received/ and triggers/)
- Update USE_CASE_MAP imports; remove stale SEMANTICS_HANDLERS alias
- Update FastAPI actors.py router import for action_rules
- Mirror test layout: test/core/use_cases/received/ and query/
- Add README.md documenting trigger→received→sync information flow
  per UC-ORG-04-001

Tests: 1027 passed, 5581 subtests
