+++ test/core/use_cases/received/test_actor.py
@@ -18,7 +18,7 @@ from unittest.mock import MagicMock
 
 
 from vultron.wire.as2.vocab.objects.vulnerability_case import VulnerabilityCase
-from vultron.core.use_cases.actor import (
+from vultron.core.use_cases.received.actor import (
     SuggestActorToCaseReceivedUseCase,
     AcceptSuggestActorToCaseReceivedUseCase,
     RejectSuggestActorToCaseReceivedUseCase,
