+++ vultron/core/use_cases/query/__init__.py
@@ -1,0 +1,5 @@
+"""Query use-case classes.
+
+These use cases implement read-only operations that return computed results
+from the DataLayer without modifying state.
+"""
