+++ vultron/core/use_cases/received/__init__.py
@@ -1,0 +1,7 @@
+"""Received-message use-case classes.
+
+These use cases handle inbound ActivityStreams activities that have been
+semantically extracted and routed by the dispatcher.  Each class processes
+one type of inbound message, updating local state to reflect the sender's
+assertion.
+"""
