+++ vultron/core/use_cases/received/note.py
